+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ New Features
 - Added zernikeXYAberration to map pupil coordinates to focal plane coordinates.
 - Added hexapolar util to compute hexapolar grids.
 - Added jmin to zernikePyramid to enable j<4 plots.
+- Added imin argument to Asphere.
 
 
 Performance Improvements